+++ lua/plugins/format.lua
@@ -24,6 +24,8 @@ return {
       typescript = prettier,
       typescriptreact = prettier,
       json = prettier,
+      jsonc = prettier,
+      yaml = prettier,
     },
   },
 }